numerous other fixes and improvements
- simplify type hierarchy, merging the frozen types into the base types
- use Reversible for order-sensitive FrozenOrderedBidict.__eq__
- fix bug where bidict._on_dup_kv was not set to ON_DUP_VAL as documented
- remove "pragma: no cover" comments where we should have coverage
