+++ bidict/_bidict.py
@@ -9,11 +9,14 @@
 
 from collections import MutableMapping
 
-from ._common import BidictBase, OVERWRITE, RAISE, ON_DUP_VAL
+from ._common import OVERWRITE, RAISE, ON_DUP_VAL
+from ._base import frozenbidict
 
 
-class bidict(BidictBase):  # noqa: N801; pylint: disable=invalid-name
+class bidict(frozenbidict):  # noqa: N801; pylint: disable=invalid-name
     """Mutable bidirectional map type."""
+
+    __hash__ = None  # since this class is mutable. explicit > implicit.
 
     def __delitem__(self, key):
         """Like dict's :attr:`__delitem__`."""
@@ -116,15 +119,7 @@ class bidict(BidictBase):  # noqa: N801; pylint: disable=invalid-name
         return self[key]
 
     def update(self, *args, **kw):
-        """
-        Like :attr:`putall` with default duplication behaviors.
-
-        In particular, for :class:`bidict.bidict`,
-        *on_dup_key=OVERWRITE*, *on_dup_val=RAISE*, and *on_dup_kv=RAISE*.
-
-        For :class:`bidict.LooseBidict`,
-        *on_dup_key=OVERWRITE*, *on_dup_val=OVERWRITE*, and *on_dup_kv=OVERWRITE*.
-        """
+        """Like :attr:`putall` with default duplication behaviors."""
         self._update(False, self._on_dup_key, self._on_dup_val, self._on_dup_kv, *args, **kw)
 
     def forceupdate(self, *args, **kw):
