+++ docs/namedbidict.rst.inc
@@ -24,8 +24,8 @@ whose default value is :class:`bidict.bidict`,
 allows overriding the bidict type used as the base class,
 allowing the creation of e.g. named frozen bidicts::
 
-    >>> from bidict import FrozenBidict
-    >>> ElMap = namedbidict('ElMap', 'sym', 'el', base_type=FrozenBidict)
+    >>> from bidict import frozenbidict
+    >>> ElMap = namedbidict('ElMap', 'sym', 'el', base_type=frozenbidict)
     >>> noble = ElMap(He='helium')
     >>> hash(noble) is not 'an error'
     True
